+++ src/onegov.people/onegov/people/collections/agencies.py
@@ -14,11 +14,3 @@ class AgencyCollection(AdjacencyListCollection):
     """
 
     __listclass__ = Agency
-
-    @property
-    def roots(self):
-        """ Returns the root agencies. """
-
-        return self.query().filter(
-            self.__listclass__.parent_id.is_(None)
-        ).all()